pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yoink-bot"
version = "0.1.0"
description = "Telegram bot for downloading media from 1000+ sites via yt-dlp"
requires-python = ">=3.12"
dependencies = [
    # Telegram
    "python-telegram-bot[job-queue]==22.7",
    # Downloaders
    "yt-dlp[default,curl-cffi]",
    "yt-dlp-get-pot",
    "gallery-dl",
    # Media processing
    "pillow",
    "ffmpeg-python",
    # Config & validation
    "pydantic>=2",
    "pydantic-settings>=2",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    # i18n
    "pyyaml",
    # Dashboard
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2",
    "python-multipart>=0.0.20",
    # API
    "python-jose[cryptography]>=3.3",
    "scalar-fastapi>=1.0",
    # Utilities
    "tldextract",
    "aiohttp",
    "aiofiles",
    "psutil",
    "sdnotify",
    # PO token provider
    "bgutil-ytdlp-pot-provider",
]

[project.optional-dependencies]
test = [
    "aiosqlite",
]

[project.scripts]
yoink-bot = "yoink.main:main"
yoink-api = "yoink.api.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
filterwarnings = ["ignore::telegram.warnings.PTBDeprecationWarning"]

[tool.uv]
compile-bytecode = false

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.9",
    "mypy>=1.13",
    "types-pyyaml>=6.0",
]
